split Backup makers into their own files under /shf_backup_makers
